updated to add some examples that use different types, removed extras
The previous author added a menu and some other stuff to actually
create a working example.  This example is less interesting, but
showcases a few extra types like a fake 'matrix' type and an int
array 'int_ptr' type that both allow for variable sizes.

Also removes the menu.h file because I'm not using it any more.